Fix for EG-341, reviewed as ENG-190, changes as follows:
This change addresses the issue raised in ENG-341.  When trying to register a ruby script through Photon, the registration process failed and the following stack-trace 
appeared in the log (partially elided):
!MESSAGE Unhandled event loop exception
!STACK 0
java.lang.VerifyError: (class: org/marketcetera/strategy/StrategyModule, method: cancelAllMarketDataRequests signature:
()V) Attempt to split long or double on the stack

The source of the problem is in JRuby which has a bug in its bindings of scalar types > 32 bits (includes doubles and floats, but not BigDecimals or int) if the types in 
question are used in the method signature of a method which was originally declared in an interface (what?!?! are you kidding?).  The problem is fixed in a later version of 
JRuby, but we cannot upgrade from 1.1.3 without some significant changes in how Ruby scripts are invoked.  We cannot make those changes now, so, instead, the fix I 
implemented was to remove the use of Java "long" types for request identifiers.  Instead, I used "int" types.  The effective difference is that users are limited to making 
2,147,483,647 simultaneous data requests across all strategies in a single JVM session instead of 9,223,372,036,854,775,807.  I strongly suspect that this will be adequate.



git-svn-id: file:///space/source.marketcetera.org@10192 501a4a96-914b-0410-8454-f5988a4ca596

+++ root/trunk/source/modules/strategy/src/main/java/org/marketcetera/strategy/ruby/Strategy.java
@@ -274,10 +274,10 @@ public class Strategy
      * 
      * @param inSymbols a <code>String</code> value containing a comma-separated list of symbols for which to request data
      * @param inSource a <code>String</code> value indicating what market data provider from which to request the data
-     * @return a <code>long</code> value containing an identifier corresponding to this market data request or 0 if the request failed
+     * @return an <code>int</code> value containing an identifier corresponding to this market data request or 0 if the request failed
      */
-    public final long request_market_data(String inSymbols,
-                                          String inSource)
+    public final int request_market_data(String inSymbols,
+                                         String inSource)
     {
         return requestMarketData(inSymbols,
                                  inSource);
@@ -292,12 +292,12 @@ public class Strategy
      *   event processor that handles them.
      * @param inCepSource a <code>String</code> value containing the name of the complex event processor
      *   to which to send the query request
-     * @return a <code>long</code> value containing the handle of the request or 0 if the request failed
+     * @return an <code>int</code> value containing the handle of the request or 0 if the request failed
      */
-    public final long request_processed_market_data(String inSymbols,
-                                                    String inMarketDataSource,
-                                                    String[] inStatements,
-                                                    String inCepSource)
+    public final int request_processed_market_data(String inSymbols,
+                                                   String inMarketDataSource,
+                                                   String[] inStatements,
+                                                   String inCepSource)
     {
         return requestProcessedMarketData(inSymbols,
                                           inMarketDataSource,
@@ -310,9 +310,9 @@ public class Strategy
      * <p>If the given <code>inRequestID</code> identifier does not correspond to an active market data
      * request, this method does nothing.
      *
-     * @param inRequestID a <code>long</code> value identifying the market data request to cancel
+     * @param inRequestID an <code>int</code> value identifying the market data request to cancel
      */
-    public final void cancel_market_data_request(long inRequestID)
+    public final void cancel_market_data_request(int inRequestID)
     {
         cancelMarketDataRequest(inRequestID);
     }
@@ -330,10 +330,10 @@ public class Strategy
      * 
      * @param inStatements a <code>String[]</code> value containing an array of statements that comprises the request
      * @param inSource a <code>String</code> value indicating what market data provider from which to request the data
-     * @return a <code>long</code> value containing an identifier corresponding to this market data request or 0 if the request failed
+     * @return an <code>int</code> value containing an identifier corresponding to this market data request or 0 if the request failed
      */
-    public final long request_cep_data(String[] inStatements,
-                                       String inSource)
+    public final int request_cep_data(String[] inStatements,
+                                      String inSource)
     {
         return requestCEPData(inStatements,
                               inSource);
@@ -344,9 +344,9 @@ public class Strategy
      * <p>If the given <code>inRequestID</code> identifier does not correspond to an active complex event processor data
      * request, this method does nothing.
      *
-     * @param inRequestID a <code>long</code> value identifying the complex event processor data request to cancel
+     * @param inRequestID an <code>int</code> value identifying the complex event processor data request to cancel
      */
-    public final void cancel_cep_request(long inRequestID)
+    public final void cancel_cep_request(int inRequestID)
     {
         cancelCEPRequest(inRequestID);
     }
